+++ main.go
@@ -8,13 +8,14 @@ import (
 	"os"
 	"strings"
 	"time"
+	"github.com/dogaakcinar/ball-possesion-rate/game"
 )
 
 func main() {
 	// Create two teams
 	var team1Time time.Duration
 	var team2Time time.Duration
-	teamNames, _ := startGame()
+	teamNames, _ := game.startGame()
 	team1 := NewTeam(teamNames[0])
 	team2 := NewTeam(teamNames[1])
 	// The starting team has the kickoff possession
@@ -50,8 +51,8 @@ func main() {
 		// Calculate the elapsed time
 
 		// Calculate possession rates based on elapsed time
-		team1.PossessionRate = calculatePossessionRate(team1Time, team1Time+team2Time)
-		team2.PossessionRate = calculatePossessionRate(team2Time, team1Time+team2Time)
+		team1.PossessionRate = game.calculatePossessionRate(team1Time, team1Time+team2Time)
+		team2.PossessionRate = game.calculatePossessionRate(team2Time, team1Time+team2Time)
 	}
 	fmt.Printf("Match summary: \n"+
 		"Total Time = %.2f \n"+
